requirements.txt
absl-py>=1.4.0
clu>=0.0.12
gin-config>=0.5.0
numpy>=2.1.3
jax>=0.6.1
flax>=0.10.6
optax>=0.2.4
orbax>=0.1.9
sentencepiece>=0.2.0
seqio>=0.0.19
tensorflow>=2.19.0
tensorflow-datasets>=4.9.8